Improve matching logic
A search token only contributes to the total score if it has a
reasonably close match with at least one of the db tokens. Additionally,
we now normalize by dividing by the number of tokens (with qualifiers
counting as half) of the larger of the searchTerm  and dbCandidate,
avoiding one-word search terms matching perfectly with any dbCandidate
that has that word in it, as normalization by max possible score
previously allowed.