Ignore filler/grammatical words in search algorithm
This prevents words like "and", "of", "for", etc. from being part of the
search criteria, except for in the exact match search, in which they are
still used. Specifically, filler words are now ignored in the candidate
reduction for the searchTerm, and the token calculation step for the
searchTerm and the dbCandidate.

ALSO note that the implementation of the token containers went from an
Array to a List.