Add additional logic for rewarding raw and penalizing processed terms
More specifically, if the database item is raw and the search term is not asking for a processed item (i.e. it explicitly states raw or doesn't say anything) then incentivize this db term by giving points.  If the database item is processed and we aren't looking explicitly for processed items in search term, decentivize this by removing points.